Update README_template.md.md
Work in lesson on ReadMe file. I do not understand this assignment at this time but am trying to learn. There is a lot of information being thrown at us and im about 2 labs behind where i should be.

The ReadMe file is being updated using Webex and postman. the response are documented although its not clear and i keep getting error responses.