Finished initial program that works
As of posting the ISS is currently flying over water so testing is difficult to see if the message displays correctly. I added some error coding for this as I would just get a callback from LocationIQ saying that it cant display the message. This is because it has no street name in the middle of the ocean.

After researching where the ISS is it seems like the locationIQ and reverse geocoding is working as the space station is in the ocean off the coast of Australia. I will come back to this tomorrow where it looks like its going to be going over Canada so will be able to test properly before moving onto commenting and extended features.